libm: more 32-bit armv8 optimization.
There's no good reason to have the floor() trick not apply to ceil(),
rint(), and trunc() (and their float variants).

Rather than duplicate the existing trick three more times, switch to a less verbose alternative that conditionally builds the msun files from the .bp file rather than via #include from builtins.cpp.

It's unlikely we'll be rid of armv7 before we stop supporting LP32
completely, so this is probably as tidy as it gets for now.

Test: treehugger
Change-Id: I746491e494d7d0cc3015cde0ec5829d7b654fdd7
Signed-off-by: Akash Srivastava <akashniki@gmail.com>